feat e refactor: adequação da página de minhasCompras para os pedidos pós venda
