+++ src/main/webapp/paginas/cliente/minhasCompras/minhasCompras.html
@@ -17,6 +17,12 @@
         <div id="container-compras" class="container mt-4"></div>
     </div>
 
+    <div class="container my-5">
+        <h1 class="mb-4 text-center">Pedidos de Troca e Devolução</h1>
+
+        <div id="container-pedidos-pos-venda" class="container mt-4"></div>
+    </div>
+
     <footer id="main-footer"></footer>
 
     <script type="module" src="js/loadPage.js"></script>
